Refactor ApiRoot to separate 'id' from 'title'
This commit refactors the ApiRoot resource to separate 'id' from 'title'.
The first implementation mistakenly used 'title' as the api root's id.
The 'id' is the actual endpoint used in referencing the ApiRoot resource.
The title is separate from the 'id', and is a human readable string.

An ApiRoot resource can be located at https://../api1, but have a title
of "Title A".