Add StatusResource21
The TAXII 2.1 Status Resource includes a modification to the fields
'successes', 'pendings', and 'failures'. They are each now a type of
'StatusDetails', which has the keys 'id', 'version', and 'message'.

'id' is the STIX Object the detail is associated with
'version' is the object's version ('modified' or 'created')
'message' indicates more information regarding the status of the object